+++ packages/messaging/client/src/client.ts
@@ -1,10 +1,9 @@
-import { uuid } from '@botpress/base'
+import { ProtectedEmitter, uuid } from '@botpress/base'
 import { Conversation, Endpoint, HealthReport, Message, SyncRequest, SyncResult, User } from '@botpress/messaging-base'
 import { AxiosRequestConfig } from 'axios'
 import { Router } from 'express'
 import { MessageFeedbackEvent } from '.'
 import { MessagingChannel } from './channel'
-import { ProtectedEmitter } from './emitter'
 import { ConversationStartedEvent, MessageNewEvent, UserNewEvent } from './events'
 import { Logger } from './logger'
 import { MessagingOptions } from './options'
