+++ packages/messaging/socket/test/tsconfig.json
@@ -3,6 +3,7 @@
   "compilerOptions": {
     "rootDir": "..",
     "paths": {
+      "@botpress/base": ["../../../base/base/src/index.ts"],
       "@botpress/messaging-base": ["../../base/src/index.ts"],
       "@botpress/base-test": ["../../../base/test/src/index.ts"]
     }
